+++ lsp/lsp-harness/src/main.rs
@@ -5,7 +5,8 @@ use lsp_types::{Position, Url};
 
 pub fn main() -> Result<()> {
     env_logger::init();
-    let s = Server::new("/home/jneeman/tweag/nickel/target/debug/nls")?;
+    let cmd = std::process::Command::new("/home/jneeman/tweag/nickel/target/debug/nls");
+    let s = Server::new(cmd)?;
     let uri = Url::parse("file:///test.ncl").unwrap();
     s.send_file(uri.clone(), "let var = 5 in { val = var }")?;
     dbg!(s.goto_def(
